Move findDocumentation implementation elsewhere
It was defined in completion.cc, but also used in hover.cc. That's a bad
practice in my opinion.
